requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
filelock==3.0.12
fsspec==2021.6.1
future==0.18.2
google-auth==1.32.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
huggingface-hub==0.0.12
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
Markdown==3.3.4
multidict==5.1.0
numpy==1.21.0
oauthlib==3.1.1
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
pytorch-lightning==1.3.7.post0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.3.2
tqdm==4.61.1
transformers==4.8.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.4.1
fairscale==0.3.7